requirements.txt
email-validator>=2.0.0
flask>=2.2.0
flask-login>=0.6.2
flask-sqlalchemy>=3.0.0
flask-wtf>=1.1.1
gunicorn>=20.1.0
matplotlib>=3.7.0
pandas>=2.0.0
psycopg2-binary>=2.9.5
routes>=2.5.1
sqlalchemy>=2.0.0
werkzeug>=2.2.0
weasyprint
weasyprint
python-dotenv
python-dotenv>=0.21.0
weasyprint>=54.0
flask-mail
flask-mail
python-telegram-bot>=20.0a1
requests>=2.28.1
APScheduler>=3.9.1
